agui_maf_demo/requirements.txt
# AG-UI with Microsoft Agent Framework Demo
# Supports both Azure OpenAI and Azure AI Foundry

# Core Agent Framework with AG-UI integration
agent-framework-ag-ui>=0.1.0

# Load environment variables from .env file
python-dotenv>=1.0.0

# Timezone support for get_current_time tool
pytz>=2023.3

# Azure AI Inference SDK (for Azure AI Foundry support)
azure-ai-inference>=1.0.0b1

# Azure Identity (for authentication)
azure-identity>=1.15.0

# Dependencies (automatically installed with agent-framework-ag-ui):
# - agent-framework-core
# - fastapi
# - uvicorn
# - httpx

# Optional: If you want to specify versions explicitly
# agent-framework-core>=0.1.0
# fastapi>=0.100.0
# uvicorn>=0.23.0
